fix(app): CI-green + UX batch — clippy cast, kotlin smart-cast, QR overflow, APK arm64-only, large limits, DMG install note, popup scroll-reset, private-mode tray resync
- core(items.rs): drop redundant usize cast (clippy -D warnings)
- android(HistoryActivity): capture produceState delegate into local val (smart-cast)
- ui(DevicesView): constrain QR svg to its 190px box (overflow)
- android(build.gradle): abiFilters + cargo targets -> arm64-v8a only (~6-8MB)
- core(defaults): large limits (history 100k, img 64MiB, file 1GiB, quota 10GiB), image_quality 100, sync_ttl 30d
- scripts(make_dmg): READ ME install note + CI-safe Finder layout
- ui(Popup): reset scrollTop+selectedIdx on hide so re-show starts at top
- ui(lib.rs tray): private-mode resync needs 2 consecutive identical IPC replies (avoid stale read from dying old daemon)

+++ crates/copypaste-core/src/storage/items.rs
@@ -1984,7 +1984,7 @@ mod tests {
         // Run prune_to_cap.
         let deleted = prune_to_cap(&db, quota).unwrap();
         assert_eq!(
-            deleted as usize,
+            deleted,
             naive_delete.len(),
             "window-fn and naive must delete the same number of rows"
         );
